Extract create/create.go into multiple files
Preceeding a larger refactoring, I'm extracting the source code of
create/create.go into multiple files.
