+++ .gitignore
@@ -1,4 +1,5 @@
 shopify-extensions
+shopify-extensions.exe
 node_modules
 __debug_bin
 .DS_Store
